tob-api/requirements.txt
Django>=1.11.5,<2
djangorestframework>=3.6.4,<4
djangorestframework-bulk>=0.2.1,<1
django-cors-headers>=2.1.0,<3

# Database managment
psycopg2>=2.7.3.1,<3

# Web Server
gunicorn>=19.7.1,<20

# Documentation
django_rest_swagger>=2.1.2,<3
